EDU-1057 Replace consent drawer with modal dialog
