+++ src/components/consent-dialog.yml
@@ -1,0 +1,6 @@
+title:
+  en: This website uses cookies
+  de: Diese Website verwendet Cookies
+acceptButton:
+  en: Accept
+  de: Akzeptieren
