--- src/components/consent-drawer.less
@@ -1,73 +1,0 @@
-@consent-drawer-padding: 15px 25px;
-
-.ConsentDrawer {
-  display: none;
-
-  &.is-enabled {
-    display: block;
-  }
-}
-
-.ConsentDrawer-overlay {
-  position: fixed;
-  left: 0;
-  top: 0;
-  width: 100%;
-  height: 100%;
-  z-index: @edu-z-index-overlay-layer;
-  background-color: rgba(0, 0, 0, 0.45);
-}
-
-.ConsentDrawer-drawer {
-  background: fade(@edu-warning-color, 85%);
-  display: flex;
-  flex-direction: row;
-  align-items: stretch;
-  justify-content: space-between;
-  position: fixed;
-  left: 0;
-  bottom: 0;
-  width: 100%;
-  z-index: @edu-z-index-sticky-content-layer;
-
-  @media @edu-media-phone {
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-  }
-}
-
-.ConsentDrawer-icon {
-  flex: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: @consent-drawer-padding;
-  font-size: 42px;
-
-  @media @edu-media-phone {
-    display: none;
-  }
-}
-
-.ConsentDrawer-consentText {
-  flex: 1 0 0%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: @consent-drawer-padding;
-  font-weight: bold;
-
-  @media @edu-media-phone {
-    text-align: center;
-    padding-bottom: 0;
-  }
-}
-
-.ConsentDrawer-acceptButton {
-  flex: none;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  padding: @consent-drawer-padding;
-}
