+++ src/styles/main.less
@@ -55,7 +55,6 @@
 @import '../components/card-selector.less';
 @import '../components/action-button.less';
 @import '../components/comments-panel.less';
-@import '../components/consent-drawer.less';
 @import '../components/credits-footer.less';
 @import '../components/pages/register.less';
 @import '../components/markdown-input.less';
